+++ logservice/logpuller/span_registry.go
@@ -16,20 +16,141 @@ package logpuller
 import (
 	"context"
 	"sync"
+	"sync/atomic"
 	"time"
 
 	"github.com/pingcap/log"
+	"github.com/pingcap/ticdc/heartbeatpb"
+	"github.com/pingcap/ticdc/logservice/logpuller/regionlock"
+	"github.com/pingcap/ticdc/pkg/common"
 	"github.com/pingcap/ticdc/pkg/metrics"
+	"github.com/pingcap/ticdc/pkg/util"
 	"github.com/tikv/client-go/v2/oracle"
 	"go.uber.org/zap"
 	"golang.org/x/sync/errgroup"
 )
+
+const kvEventsCacheMaxSize = 32
+
+// subscribedSpan is the local state for one subscribed table span.
+type subscribedSpan struct {
+	subID   SubscriptionID
+	startTs uint64
+	// Whether to filter out the value written by TiCDC itself.
+	// It should be `true` in BDR mode.
+	filterLoop bool
+
+	// The target span.
+	span heartbeatpb.TableSpan
+	// The range lock of the span,
+	// it is used to prevent duplicate requests to the same region range,
+	// and it also used to calculate this table's resolvedTs.
+	rangeLock *regionlock.RangeLock
+
+	consumeKVEvents func(events []common.RawKVEntry, wakeCallback func()) bool
+
+	advanceResolvedTs func(ts uint64)
+
+	advanceInterval int64
+
+	kvEventsCache []common.RawKVEntry
+
+	// To handle span removing.
+	stopped atomic.Bool
+
+	// To handle stale lock resolvings.
+	tryResolveLock     func(regionID uint64, state *regionlock.LockedRangeState)
+	staleLocksTargetTs atomic.Uint64
+
+	lastAdvanceTime atomic.Int64
+
+	initialized       atomic.Bool
+	resolvedTsUpdated atomic.Int64
+	resolvedTs        atomic.Uint64
+}
 
 type spanRegistry struct {
 	sync.RWMutex
 	spans map[SubscriptionID]*subscribedSpan
 
 	upstream *upstreamHandle
+}
+
+type spanAndTargetTs struct {
+	span     *subscribedSpan
+	targetTs uint64
+}
+
+func newSubscribedSpan(
+	ctx context.Context,
+	resolveLockRateLimiter *resolveLockRateLimiter,
+	resolveLockTaskCh chan resolveLockTask,
+	subID SubscriptionID,
+	span heartbeatpb.TableSpan,
+	startTs uint64,
+	consumeKVEvents func(raw []common.RawKVEntry, wakeCallback func()) bool,
+	advanceResolvedTs func(ts uint64),
+	advanceInterval int64,
+	filterLoop bool,
+) *subscribedSpan {
+	rangeLock := regionlock.NewRangeLock(uint64(subID), span.StartKey, span.EndKey, startTs)
+
+	rt := &subscribedSpan{
+		subID:      subID,
+		span:       span,
+		startTs:    startTs,
+		filterLoop: filterLoop,
+		rangeLock:  rangeLock,
+
+		consumeKVEvents:   consumeKVEvents,
+		advanceResolvedTs: advanceResolvedTs,
+		advanceInterval:   advanceInterval,
+	}
+	rt.initialized.Store(false)
+	rt.resolvedTsUpdated.Store(time.Now().Unix())
+	rt.resolvedTs.Store(startTs)
+
+	rt.tryResolveLock = func(regionID uint64, state *regionlock.LockedRangeState) {
+		targetTs := rt.staleLocksTargetTs.Load()
+		if !state.Initialized.Load() || state.ResolvedTs.Load() >= targetTs {
+			return
+		}
+		key := resolveLockKey{keyspaceID: span.KeyspaceID, regionID: regionID}
+		if !resolveLockRateLimiter.trySchedule(key, time.Now()) {
+			return
+		}
+		select {
+		case <-ctx.Done():
+			resolveLockRateLimiter.cancel(key)
+		case resolveLockTaskCh <- resolveLockTask{
+			keyspaceID: span.KeyspaceID,
+			regionID:   regionID,
+			targetTs:   targetTs,
+			state:      state,
+		}:
+		// it is ok to ignore resolve lock task when the channel is full
+		default:
+			resolveLockRateLimiter.cancel(key)
+			metrics.SubscriptionClientResolveLockTaskDropCounter.Inc()
+		}
+	}
+	return rt
+}
+
+func (span *subscribedSpan) clearKVEventsCache() {
+	if cap(span.kvEventsCache) > kvEventsCacheMaxSize {
+		span.kvEventsCache = nil
+	} else {
+		span.kvEventsCache = span.kvEventsCache[:0]
+	}
+}
+
+func (span *subscribedSpan) resolveStaleLocks(targetTs uint64) {
+	util.MustCompareAndMonotonicIncrease(&span.staleLocksTargetTs, targetTs)
+	res := span.rangeLock.IterAll(span.tryResolveLock)
+	log.Debug("subscription client finds slow locked ranges",
+		zap.Uint64("subscriptionID", uint64(span.subID)),
+		zap.Any("ranges", res))
 }
 
 func newSpanRegistry(upstream *upstreamHandle) *spanRegistry {
@@ -93,7 +214,7 @@ func (r *spanRegistry) runResolveLockChecker(ctx context.Context) error {
 	resolveLockTicker := time.NewTicker(resolveLockTickInterval)
 	defer resolveLockTicker.Stop()
 	maxCacheSize := 1024
-	subSpanAndTsCache := make([]subscriptionAndTargetTs, 0, maxCacheSize)
+	spanAndTsCache := make([]spanAndTargetTs, 0, maxCacheSize)
 	// getResolvedTargetTs returns the targetTs to resolve stale locks. 0 means no need to resolve.
 	getResolvedTargetTs := func(subSpan *subscribedSpan, currentTime time.Time, currentTs uint64) uint64 {
 		resolvedTsUpdated := time.Unix(subSpan.resolvedTsUpdated.Load(), 0)
@@ -127,20 +248,20 @@ func (r *spanRegistry) runResolveLockChecker(ctx context.Context) error {
 			if subSpan != nil {
 				targetTs := getResolvedTargetTs(subSpan, currentTime, currentTs)
 				if targetTs > 0 {
-					subSpanAndTsCache = append(subSpanAndTsCache, subscriptionAndTargetTs{
-						subSpan:  subSpan,
+					spanAndTsCache = append(spanAndTsCache, spanAndTargetTs{
+						span:     subSpan,
 						targetTs: targetTs,
 					})
 				}
 			}
 		}
 		r.RUnlock()
-		for _, subSpanAndTs := range subSpanAndTsCache {
-			subSpanAndTs.subSpan.resolveStaleLocks(subSpanAndTs.targetTs)
+		for _, spanAndTs := range spanAndTsCache {
+			spanAndTs.span.resolveStaleLocks(spanAndTs.targetTs)
 		}
-		subSpanAndTsCache = subSpanAndTsCache[:0]
-		if cap(subSpanAndTsCache) > maxCacheSize {
-			subSpanAndTsCache = make([]subscriptionAndTargetTs, 0, maxCacheSize)
+		spanAndTsCache = spanAndTsCache[:0]
+		if cap(spanAndTsCache) > maxCacheSize {
+			spanAndTsCache = make([]spanAndTargetTs, 0, maxCacheSize)
 		}
 	}
 }
